DTFS2-4456 - split out parties controller functions
