+++ trade-finance-manager-ui/server/controllers/case/index.test.js
@@ -646,122 +646,6 @@ describe('controllers - case', () => {
     });
   });
 
-
-  describe('GET case parties', () => {
-    describe('when deal exists', () => {
-      const mockDeal = {
-        _id: '1000023',
-        dealSnapshot: {
-          _id: '1000023',
-        },
-        mock: true,
-      };
-
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve(mockDeal);
-      });
-
-      it('should render deal template with data', async () => {
-        const req = {
-          params: {
-            _id: mockDeal._id,
-          },
-          session,
-        };
-
-        await caseController.getCaseParties(req, res);
-        expect(res.render).toHaveBeenCalledWith('case/parties/parties.njk', {
-          deal: mockDeal.dealSnapshot,
-          activePrimaryNavigation: 'manage work',
-          activeSubNavigation: 'parties',
-          dealId: req.params._id,
-          user: session.user,
-        });
-      });
-    });
-
-    describe('when deal does NOT exist', () => {
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve();
-      });
-
-      it('should redirect to not-found route', async () => {
-        const req = {
-          params: {
-            _id: '1',
-          },
-          session,
-        };
-
-        await caseController.getCaseParties(req, res);
-        expect(res.redirect).toHaveBeenCalledWith('/not-found');
-      });
-    });
-  });
-
-  describe('GET party edit', () => {
-    describe('when deal exists', () => {
-      const mockDeal = {
-        _id: '1000023',
-        dealSnapshot: {
-          _id: '1000023',
-          submissionDetails: {
-            supplierName: 'test supplier',
-          },
-        },
-        tfm: {
-          parties: {
-            exporter: {
-              partyUrn: '12345',
-            },
-          },
-        },
-        mock: true,
-      };
-
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve(mockDeal);
-      });
-
-      it('should render edit template with data', async () => {
-        const req = {
-          params: {
-            _id: mockDeal._id,
-          },
-          session,
-        };
-
-        await caseController.getExporterPartyDetails(req, res);
-        expect(res.render).toHaveBeenCalledWith('case/parties/edit/exporter-edit.njk', {
-          activePrimaryNavigation: 'manage work',
-          activeSubNavigation: 'parties',
-          deal: mockDeal.dealSnapshot,
-          tfm: mockDeal.tfm,
-          dealId: req.params._id,
-          user: session.user,
-        });
-      });
-    });
-
-    describe('when deal does NOT exist', () => {
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve();
-      });
-
-      it('should redirect to not-found route', async () => {
-        const req = {
-          params: {
-            _id: '1',
-          },
-          session,
-        };
-
-        await caseController.getExporterPartyDetails(req, res);
-        expect(res.redirect).toHaveBeenCalledWith('/not-found');
-      });
-    });
-  });
-
   describe('GET case documents', () => {
     describe('when deal exists', () => {
       const mockDeal = {
@@ -816,132 +700,7 @@ describe('controllers - case', () => {
     });
   });
 
-  describe('GET bond issuer edit', () => {
-    describe('when deal exists', () => {
-      const mockDeal = {
-        _id: '1000023',
-        dealSnapshot: {
-          _id: '1000023',
-          submissionDetails: {
-            supplierName: 'test supplier',
-          },
-        },
-        tfm: {
-          parties: {
-            exporter: {
-              partyUrn: '12345',
-            },
-          },
-        },
-        mock: true,
-      };
-
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve(mockDeal);
-      });
-
-      it('should render bond edit template with data', async () => {
-        const req = {
-          params: {
-            _id: mockDeal._id,
-          },
-          session,
-        };
-
-        await caseController.getBondIssuerPartyDetails(req, res);
-        expect(res.render).toHaveBeenCalledWith('case/parties/edit/bonds-issuer-edit.njk',
-          {
-            activePrimaryNavigation: 'manage work',
-            activeSubNavigation: 'parties',
-            deal: mockDeal.dealSnapshot,
-            user: session.user,
-          });
-      });
-    });
-
-    describe('when deal does NOT exist', () => {
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve();
-      });
-
-      it('should redirect to not-found route', async () => {
-        const req = {
-          params: {
-            _id: '1',
-          },
-          session,
-        };
-
-        await caseController.getBondIssuerPartyDetails(req, res);
-        expect(res.redirect).toHaveBeenCalledWith('/not-found');
-      });
-    });
-  });
-
-  describe('GET bond beneficiary edit', () => {
-    describe('when deal exists', () => {
-      const mockDeal = {
-        _id: '1000023',
-        dealSnapshot: {
-          _id: '1000023',
-          submissionDetails: {
-            supplierName: 'test supplier',
-          },
-        },
-        tfm: {
-          parties: {
-            exporter: {
-              partyUrn: '12345',
-            },
-          },
-        },
-        mock: true,
-      };
-
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve(mockDeal);
-      });
-
-      it('should render bond edit template with data', async () => {
-        const req = {
-          params: {
-            _id: mockDeal._id,
-          },
-          session,
-        };
-
-        await caseController.getBondBeneficiaryPartyDetails(req, res);
-        expect(res.render).toHaveBeenCalledWith('case/parties/edit/bonds-beneficiary-edit.njk',
-          {
-            activePrimaryNavigation: 'manage work',
-            activeSubNavigation: 'parties',
-            deal: mockDeal.dealSnapshot,
-            user: session.user,
-          });
-      });
-    });
-
-    describe('when deal does NOT exist', () => {
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve();
-      });
-
-      it('should redirect to not-found route', async () => {
-        const req = {
-          params: {
-            _id: '1',
-          },
-          session,
-        };
-
-        await caseController.getBondBeneficiaryPartyDetails(req, res);
-        expect(res.redirect).toHaveBeenCalledWith('/not-found');
-      });
-    });
-  });
-
-
-  describe('POST party edit', () => {
+  describe('POST facility edit', () => {
     describe('when deal exists', () => {
       const mockDeal = {
         _id: '1000023',
@@ -952,56 +711,7 @@ describe('controllers - case', () => {
         api.getDeal = () => Promise.resolve(mockDeal);
       });
 
-      it('should render bond edit template with data', async () => {
-        const req = {
-          params: {
-            _id: mockDeal._id,
-          },
-          body: {
-            exporter: {
-              partyUrn: '12345',
-            },
-          },
-          session,
-        };
-
-        await caseController.postExporterPartyDetails(req, res);
-        // eslint-disable-next-line no-underscore-dangle
-        expect(res.redirect).toHaveBeenCalledWith(`/case/${mockDeal._id}/parties`);
-      });
-    });
-
-    describe('when deal does NOT exist', () => {
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve();
-      });
-
-      it('should redirect to not-found route', async () => {
-        const req = {
-          params: {
-            _id: '1',
-          },
-          session,
-        };
-
-        await caseController.postExporterPartyDetails(req, res);
-        expect(res.redirect).toHaveBeenCalledWith('/not-found');
-      });
-    });
-  });
-
-  describe('POST bond edit', () => {
-    describe('when deal exists', () => {
-      const mockDeal = {
-        _id: '1000023',
-        mock: true,
-      };
-
-      beforeEach(() => {
-        api.getDeal = () => Promise.resolve(mockDeal);
-      });
-
-      it('should render bond edit template with data', async () => {
+      it('should redirect to parties page', async () => {
         const req = {
           params: {
             _id: mockDeal._id,
